Add health check runtime test for memories, fix fram driver and fix internal flash device driver

+++ firmware/app/tasks/read_px.c
@@ -24,14 +24,17 @@
  * \brief Read PX task implementation.
  *
  * \author Augusto Cezar Boldori Vassoler <augustovassoler@gmail.com>
- *
- * \version 0.0.1
+ * \author Carlos Augusto Porto Freitas <carlos.portof@hotmail.com>
+ * 
+ * \version 0.10.18
  *
  * \date 2023/08/28
  *
  * \addtogroup read_px
  * \{
  */
+
+#include <stdint.h>
 
 #include <system/sys_log/sys_log.h>
 #include <devices/payload/payload.h>
@@ -61,8 +64,8 @@ void vTaskReadPX(void)
         {
             sys_log_print_event_from_module(SYS_LOG_ERROR, TASK_READ_PX_NAME, "Error reading the ping-pong data!");
 
-            uint8_t i = 0;
-            for(i=0;i<px_buf.length;i++){
+            int32_t i = 0;
+            for(i = 0; i < px_buf.length; i++){
                 sys_log_print_uint(px_buf.buffer[i]);
             }
